solution: splitting classes into methods and demo for theme oop-principles
