+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task1.java
@@ -1,7 +1,6 @@
 package com.rakovets.course.java.core.practice.exception_handling;
 
 public class Task1 {
-
     public static void main(String[] args) {
         String string = null;
