+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task2.java
@@ -1,7 +1,6 @@
 package com.rakovets.course.java.core.practice.exception_handling;
 
 public class Task2 {
-
     public static void main(String[] args) {
         String[] string = new String[3];
