+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task5.java
@@ -1,7 +1,6 @@
 package com.rakovets.course.java.core.practice.exception_handling;
 
 public class Task5 {
-
     public static void arithmetic2 (String x) throws CustomRunTimeException {
         int y;
         try {
