+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task6.java
@@ -1,7 +1,6 @@
 package com.rakovets.course.java.core.practice.exception_handling;
 
 public class Task6 {
-
     public static void randomException() throws ArithmeticException, NumberFormatException, IllegalArgumentException {
         int x = (int) (Math.random() * 3);
 
@@ -13,16 +12,6 @@ public class Task6 {
         }
         else {
             throw new IllegalArgumentException();
-        }
-    }
-
-    public static void main(String[] args) {
-        try {
-            randomException();
-        } catch (ArithmeticException | NumberFormatException ex) {
-            System.out.println("Arithmetic or Number exceptions are caught");
-        } catch (IllegalArgumentException ill) {
-            System.out.println("IllegalArgumentException caught");
         }
     }
 }
