+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task7.java
@@ -1,7 +1,6 @@
 package com.rakovets.course.java.core.practice.exception_handling;
 
 public class Task7 {
-
     public static void fiftyPercentChance() throws ArithmeticException, NumberFormatException {
         int x = (int) (Math.random() * 2);
 
@@ -9,18 +8,6 @@ public class Task7 {
             throw new ArithmeticException();
         } else {
             throw new IllegalArgumentException();
-        }
-    }
-
-    public static void main(String[] args) {
-        try {
-            fiftyPercentChance();
-        } catch (ArithmeticException a) {
-            System.out.println("ArithmeticException caught");
-        } catch (IllegalArgumentException i) {
-            System.out.println("IllegalArgumentException caugth");
-        } finally {
-            System.out.println("Program is finished");
         }
     }
 }
