+++ src/main/java/com/rakovets/course/java/core/practice/generic_types/Math.java
@@ -1,0 +1,5 @@
+package com.rakovets.course.java.core.practice.generic_types;
+
+public abstract class Math {
+
+}
